Add nexus-staging plugin to "plugins" section
Before this change, running "mvn deploy" would result in running the
"maven-deploy-plugin:deploy" goal. This behavior differs from Java
Driver 4.x, which executes a "nexus-staging-maven-plugin:deploy" goal.

This change should fix auto-closing of Sonatype repository when
releasing the driver, as is already done in Java Driver 4.x. This will
make it easier in the future to automate the release of Java Driver 3.x,
as a maintainer won't have to manually close Sonetype repository through
a website.